+++ jaclang/plugin/spec.py
@@ -212,7 +212,7 @@ class JacBuiltin:
     @hookspec(firstresult=True)
     def dotgen(
         node: NodeArchitype,
-        depth: float,
+        depth: int,
         traverse: bool,
         edge_type: list[str],
         bfs: bool,
